refactor(repo): scripts/lib → scripts/lint/helpers (only lint uses it)
