+++ scripts/lint/cross-file/contract-snapshots.ts
@@ -8,7 +8,7 @@
 // inside an existing fixture stay a CI-only catch.
 import { readdirSync, readFileSync } from "node:fs";
 import { relative, resolve } from "node:path";
-import { REPO_ROOT } from "../../lib/paths.ts";
+import { REPO_ROOT } from "../helpers/paths.ts";
 import type { ViolationDetail, WorkspaceCheck } from "../registry.ts";
 
 const CONTRACT_DIR = resolve(REPO_ROOT, "tests", "contract");
