+++ scripts/lint/cross-file/dogfood.ts
@@ -7,7 +7,7 @@
 // being authored. The marker logs the gap rather than hiding it.
 import { readdirSync, readFileSync } from "node:fs";
 import { join, relative, resolve } from "node:path";
-import { REPO_ROOT } from "../../lib/paths.ts";
+import { REPO_ROOT } from "../helpers/paths.ts";
 import type { ViolationDetail, WorkspaceCheck } from "../registry.ts";
 
 function walk(dir: string): string[] {
