+++ scripts/lint/cross-file/exports-resolve.ts
@@ -12,7 +12,7 @@
 // still flags real misses.
 import { existsSync, readdirSync, readFileSync, statSync } from "node:fs";
 import { dirname, resolve } from "node:path";
-import { REPO_ROOT } from "../../lib/paths.ts";
+import { REPO_ROOT } from "../helpers/paths.ts";
 import type { ViolationDetail, WorkspaceCheck } from "../registry.ts";
 
 type ExportsValue = string | null | { [condition: string]: ExportsValue };
