+++ scripts/lint/cross-file/label-sync.ts
@@ -2,7 +2,7 @@
 // does not define. The two files drift independently; this is the gate.
 import { readFileSync } from "node:fs";
 import { resolve } from "node:path";
-import { REPO_ROOT } from "../../lib/paths.ts";
+import { REPO_ROOT } from "../helpers/paths.ts";
 import type { ViolationDetail, WorkspaceCheck } from "../registry.ts";
 
 function labelerReferences(): string[] {
