+++ scripts/lint/file-rules/motion-scale.ts
@@ -4,7 +4,7 @@
 import { readdirSync, readFileSync } from "node:fs";
 import { relative, resolve } from "node:path";
 import postcss from "postcss";
-import { REPO_ROOT } from "../../lib/paths.ts";
+import { REPO_ROOT } from "../helpers/paths.ts";
 import type { ViolationDetail, WorkspaceCheck } from "../registry.ts";
 
 const MOTION_PROPS = new Set([
