+++ scripts/lint/file-rules/rhythm-tokens.ts
@@ -15,7 +15,7 @@ import { readdirSync, readFileSync } from "node:fs";
 import { relative, resolve } from "node:path";
 import postcss from "postcss";
 import postcssEach from "postcss-each";
-import { REPO_ROOT } from "../../lib/paths.ts";
+import { REPO_ROOT } from "../helpers/paths.ts";
 import type { ViolationDetail, WorkspaceCheck } from "../registry.ts";
 
 const SIZING_PROPS = new Set([
