+++ scripts/lint/helpers/enumerate.test.ts
