+++ scripts/lint/helpers/enumerate.ts
