+++ scripts/lint/helpers/glob.test.ts
