+++ scripts/lint/helpers/glob.ts
